Extract compute resource logic into separate functions
