+++ frontend/src/old-pages/Configure/Queues.tsx
@@ -35,23 +35,25 @@ import {setState, getState, useState, clearState} from '../../store'
 import {
   ActionsEditor,
   CustomAMISettings,
-  InstanceSelect,
   LabeledIcon,
   RootVolume,
   SubnetSelect,
   SecurityGroups,
   IamPoliciesEditor,
-  HelpTextInput,
 } from './Components'
-import HelpTooltip from '../../components/HelpTooltip'
 import {Trans, useTranslation} from 'react-i18next'
 import {SlurmMemorySettings} from './SlurmMemorySettings'
-import {useFeatureFlag} from '../../feature-flags/useFeatureFlag'
+import {
+  isFeatureEnabled,
+  useFeatureFlag,
+} from '../../feature-flags/useFeatureFlag'
+import * as SingleInstanceCR from './SingleInstanceComputeResource'
+import * as MultiInstanceCR from './MultiInstanceComputeResource'
+import {ComputeResource} from './queues.types'
 
 // Constants
 const queuesPath = ['app', 'wizard', 'config', 'Scheduling', 'SlurmQueues']
 const queuesErrorsPath = ['app', 'wizard', 'errors', 'queues']
-const defaultInstanceType = 'c5n.large'
 
 // Helper Functions
 // @ts-expect-error TS(7031) FIXME: Binding element 'value' implicitly has an 'any' ty... Remove this comment to see the full error message
@@ -179,19 +181,30 @@ function queueValidate(queueIndex: any) {
     setState([...errorsPath, 'subnet'], null)
   }
 
-  let seenInstances = new Set()
-  for (let i = 0; i < computeResources.length; i++) {
-    let computeResource = computeResources[i]
-    if (seenInstances.has(computeResource.InstanceType)) {
-      setState(
-        [...errorsPath, 'computeResource', i, 'type'],
-        i18next.t('wizard.queues.validation.instanceTypeUnique'),
-      )
-      valid = false
-    } else {
-      seenInstances.add(computeResource.InstanceType)
-      setState([...errorsPath, 'computeResource', i, 'type'], null)
-    }
+  const version = getState(['app', 'version', 'full'])
+  const isMultiInstanceTypesActive = isFeatureEnabled(
+    version,
+    'queues_multiple_instance_types',
+  )
+  // TODO: call multiple instance validator
+  const {validateComputeResources} = !isMultiInstanceTypesActive
+    ? SingleInstanceCR
+    : SingleInstanceCR
+  const [computeResourcesValid, computeResourcesErrors] =
+    validateComputeResources(computeResources)
+  if (!computeResourcesValid) {
+    valid = false
+    computeResources.forEach((_: ComputeResource, i: number) => {
+      const error = computeResourcesErrors[i]
+      if (error) {
+        setState(
+          [...errorsPath, 'computeResource', i, 'type'],
+          i18next.t('wizard.queues.validation.instanceTypeUnique'),
+        )
+      } else {
+        setState([...errorsPath, 'computeResource', i, 'type'], null)
+      }
+    })
   }
 
   return valid
@@ -213,254 +226,12 @@ function queuesValidate() {
   return valid
 }
 
-function ComputeResource({index, queueIndex, computeResource}: any) {
-  const parentPath = [...queuesPath, queueIndex]
-  const queue = useState(parentPath)
-  const computeResources = useState([...parentPath, 'ComputeResources'])
-  const path = [...parentPath, 'ComputeResources', index]
-  const errorsPath = [...queuesErrorsPath, queueIndex, 'computeResource', index]
-  const typeError = useState([...errorsPath, 'type'])
-
-  const tInstances = new Set<string>(['t2.micro', 't2.medium'])
-  const gravitonInstances = new Set<string>([])
-
-  const instanceTypePath = [...path, 'InstanceType']
-  const instanceType: string = useState(instanceTypePath)
-  const memoryBasedSchedulingEnabledPath = [
-    'app',
-    'wizard',
-    'config',
-    'Scheduling',
-    'SlurmSettings',
-    'EnableMemoryBasedScheduling',
-  ]
-  const enableMemoryBasedScheduling = useState(memoryBasedSchedulingEnabledPath)
-
-  const disableHTPath = [...path, 'DisableSimultaneousMultithreading']
-  const disableHT = useState(disableHTPath)
-
-  const efaPath = [...path, 'Efa']
-
-  const efaInstances = new Set(useState(['aws', 'efa_instance_types']))
-  const enableEFAPath = [...path, 'Efa', 'Enabled']
-  const enableEFA = useState(enableEFAPath) || false
-
-  const enablePlacementGroupPath = [
-    ...parentPath,
-    'Networking',
-    'PlacementGroup',
-    'Enabled',
-  ]
-
-  const enableGPUDirectPath = [...path, 'Efa', 'GdrSupport'] || false
-  const enableGPUDirect = useState(enableGPUDirectPath)
-
-  const instanceSupportsGdr = instanceType === 'p4d.24xlarge'
-
-  const minCount = useState([...path, 'MinCount'])
-  const maxCount = useState([...path, 'MaxCount'])
-
-  const {t} = useTranslation()
-
-  const remove = () => {
-    setState(
-      [...parentPath, 'ComputeResources'],
-      [
-        ...computeResources.slice(0, index),
-        ...computeResources.slice(index + 1),
-      ],
-    )
-  }
-
-  const setMinCount = (staticCount: any) => {
-    const dynamicCount = maxCount - minCount
-    if (staticCount > 0)
-      setState([...path, 'MinCount'], !isNaN(staticCount) ? staticCount : 0)
-    else clearState([...path, 'MinCount'])
-    setState(
-      [...path, 'MaxCount'],
-      (!isNaN(staticCount) ? staticCount : 0) +
-        (!isNaN(dynamicCount) ? dynamicCount : 0),
-    )
-  }
-
-  const setMaxCount = (dynamicCount: any) => {
-    const staticCount = minCount
-    setState(
-      [...path, 'MaxCount'],
-      (!isNaN(staticCount) ? staticCount : 0) +
-        (!isNaN(dynamicCount) ? dynamicCount : 0),
-    )
-  }
-
-  const setSchedulableMemory = (
-    schedulableMemoryPath: string[],
-    schedulableMemory: string,
-  ) => {
-    let schedulableMemoryNumber = parseInt(schedulableMemory)
-    if (enableMemoryBasedScheduling && !isNaN(schedulableMemoryNumber)) {
-      setState(schedulableMemoryPath, schedulableMemoryNumber)
-    } else {
-      clearState(schedulableMemoryPath)
-    }
-  }
-
-  const setDisableHT = (disable: any) => {
-    if (disable) setState(disableHTPath, disable)
-    else clearState(disableHTPath)
-  }
-
-  const setEnableEFA = (enable: any) => {
-    if (enable) {
-      setState(enableEFAPath, enable)
-      setState(enablePlacementGroupPath, enable)
-    } else {
-      clearState(efaPath)
-      clearState(enablePlacementGroupPath)
-    }
-  }
-
-  const setEnableGPUDirect = (enable: any) => {
-    if (enable) setState(enableGPUDirectPath, enable)
-    else clearState(enableGPUDirectPath)
-  }
-
-  const setInstanceType = (instanceType: any) => {
-    // setting the instance type on the queue happens in the component
-    // this updates the name which is derived from the instance type
-    setState(
-      [...path, 'Name'],
-      `${queue.Name}-${instanceType.replace('.', '')}`,
-    )
-
-    if (!getState(enableEFAPath) && efaInstances.has(instanceType))
-      setEnableEFA(true)
-    else if (getState(enableEFAPath) && !efaInstances.has(instanceType))
-      setEnableEFA(false)
-  }
-
-  React.useEffect(() => {
-    if (!instanceType)
-      setState(
-        [...queuesPath, queueIndex, 'ComputeResources', index, 'InstanceType'],
-        defaultInstanceType,
-      )
-  }, [queueIndex, index, instanceType])
-
-  return (
-    <div className="compute-resource">
-      <div style={{display: 'flex', flexDirection: 'column', gap: '10px'}}>
-        <Box margin={{top: 'xs'}} textAlign="right">
-          {index > 0 && <Button onClick={remove}>Remove Resource</Button>}
-        </Box>
-        <ColumnLayout columns={2}>
-          <div style={{display: 'flex', flexDirection: 'row', gap: '20px'}}>
-            <FormField label={t('wizard.queues.computeResource.staticNodes')}>
-              <Input
-                value={computeResource.MinCount || 0}
-                type="number"
-                onChange={({detail}) => setMinCount(parseInt(detail.value))}
-              />
-            </FormField>
-            <FormField label={t('wizard.queues.computeResource.dynamicNodes')}>
-              <Input
-                value={Math.max(
-                  (computeResource.MaxCount || 0) -
-                    (computeResource.MinCount || 0),
-                  0,
-                ).toString()}
-                type="number"
-                onChange={({detail}) => setMaxCount(parseInt(detail.value))}
-              />
-            </FormField>
-          </div>
-          <FormField
-            label={t('wizard.queues.computeResource.instanceType')}
-            errorText={typeError}
-          >
-            <InstanceSelect
-              path={instanceTypePath}
-              callback={setInstanceType}
-            />
-          </FormField>
-          {enableMemoryBasedScheduling && (
-            <HelpTextInput
-              name={t('wizard.queues.schedulableMemory.name')}
-              path={path}
-              errorsPath={errorsPath}
-              configKey={'SchedulableMemory'}
-              onChange={({detail}) =>
-                setSchedulableMemory(
-                  [...path, 'SchedulableMemory'],
-                  detail.value,
-                )
-              }
-              description={t('wizard.queues.schedulableMemory.description')}
-              placeholder={t('wizard.queues.schedulableMemory.placeholder')}
-              help={t('wizard.queues.schedulableMemory.help')}
-              type="number"
-            />
-          )}
-        </ColumnLayout>
-        <div style={{display: 'flex', flexDirection: 'row', gap: '20px'}}>
-          <Toggle
-            disabled={
-              tInstances.has(instanceType) ||
-              gravitonInstances.has(instanceType)
-            }
-            checked={disableHT}
-            onChange={_e => {
-              setDisableHT(!disableHT)
-            }}
-          >
-            <Trans i18nKey="wizard.queues.computeResource.disableHT" />
-          </Toggle>
-          <Toggle
-            disabled={!efaInstances.has(instanceType)}
-            checked={enableEFA}
-            onChange={_e => {
-              setEnableEFA(!enableEFA)
-            }}
-          >
-            <Trans i18nKey="wizard.queues.computeResource.enableEfa" />
-          </Toggle>
-          <div
-            style={{
-              display: 'flex',
-              flexDirection: 'row',
-              alignItems: 'center',
-            }}
-          >
-            <Toggle
-              disabled={!instanceSupportsGdr}
-              checked={enableGPUDirect}
-              onChange={_e => {
-                setEnableGPUDirect(!enableGPUDirect)
-              }}
-            >
-              <Trans i18nKey="wizard.queues.computeResource.enableGpuDirect" />
-            </Toggle>
-            <HelpTooltip>
-              <Trans i18nKey="wizard.queues.Gdr.help">
-                <a
-                  rel="noreferrer"
-                  target="_blank"
-                  href="https://docs.aws.amazon.com/parallelcluster/latest/ug/Scheduling-v3.html#yaml-Scheduling-SlurmQueues-ComputeResources-Efa-GdrSupport"
-                ></a>
-              </Trans>
-            </HelpTooltip>
-          </div>
-        </div>
-      </div>
-    </div>
-  )
-}
-
 function ComputeResources({queue, index}: any) {
+  const {ViewComponent} = useComputeResourceAdapter()
   return (
     <Container>
       {queue.ComputeResources.map((computeResource: any, i: any) => (
-        <ComputeResource
+        <ViewComponent
           queue={queue}
           computeResource={computeResource}
           index={i}
@@ -476,6 +247,7 @@ function Queue({index}: any) {
   const {t} = useTranslation()
   const queues = useState(queuesPath)
   const [editingName, setEditingName] = React.useState(false)
+  const computeResourceAdapter = useComputeResourceAdapter()
   const queue = useState([...queuesPath, index])
   const enablePlacementGroupPath = [
     ...queuesPath,
@@ -510,12 +282,7 @@ function Queue({index}: any) {
       ...queue,
       ComputeResources: [
         ...(queue.ComputeResources || []),
-        {
-          Name: `queue${index}-${defaultInstanceType.replace('.', '')}`,
-          InstanceType: defaultInstanceType,
-          MinCount: 0,
-          MaxCount: 4,
-        },
+        computeResourceAdapter.createComputeResource(index),
       ],
     })
   }
@@ -530,12 +297,12 @@ function Queue({index}: any) {
       index,
       'ComputeResources',
     ])
-    for (let i = 0; i < computeResources.length; i++) {
-      const cr = computeResources[i]
-      const crName = `${newName}-${cr.InstanceType.replace('.', '')}`
-      setState([...queuesPath, index, 'ComputeResources', i, 'Name'], crName)
-    }
+    const updatedCRs = computeResourceAdapter.updateComputeResourcesNames(
+      computeResources,
+      newName,
+    )
     setState([...queuesPath, index, 'Name'], newName)
+    setState([...queuesPath, index, 'ComputeResources'], updatedCRs)
   }
 
   return (
@@ -666,6 +433,7 @@ function Queues() {
   const isMemoryBasedSchedulingActive = useFeatureFlag(
     'memory_based_scheduling',
   )
+  const adapter = useComputeResourceAdapter()
   let queues = useState(queuesPath) || []
 
   const addQueue = () => {
@@ -675,17 +443,7 @@ function Queues() {
         ...(queues || []),
         {
           Name: `queue${queues.length}`,
-          ComputeResources: [
-            {
-              Name: `queue${queues.length}-${defaultInstanceType.replace(
-                '.',
-                '',
-              )}`,
-              MinCount: 0,
-              MaxCount: 4,
-              InstanceType: defaultInstanceType,
-            },
-          ],
+          ComputeResources: [adapter.createComputeResource(queues.length)],
         },
       ],
     )
@@ -714,6 +472,28 @@ function Queues() {
       </Container>
     </ColumnLayout>
   )
+}
+
+const useComputeResourceAdapter = () => {
+  const isMultiInstanceTypesActive = useFeatureFlag(
+    'queues_multiple_instance_types',
+  )
+  return !isMultiInstanceTypesActive
+    ? {
+        ViewComponent: SingleInstanceCR.ComputeResource,
+        updateComputeResourcesNames:
+          SingleInstanceCR.updateComputeResourcesNames,
+        createComputeResource: SingleInstanceCR.createComputeResource,
+        validateComputeResources: SingleInstanceCR.validateComputeResources,
+      }
+    : {
+        // TODO: implement multi instance variant
+        ViewComponent: SingleInstanceCR.ComputeResource,
+        updateComputeResourcesNames:
+          SingleInstanceCR.updateComputeResourcesNames,
+        createComputeResource: SingleInstanceCR.createComputeResource,
+        validateComputeResources: SingleInstanceCR.validateComputeResources,
+      }
 }
 
 export {Queues, queuesValidate}
